[FLINK-11485] [core] Move OptionalMap to org.apache.flink.util
This utility class may commonly be used by the KryoSerializer and the
PojoSerializer. This commit moves the utility to a more suitable package
and adjusts usage access.
